fixed bridging history
